requirements.txt
# Google Ads MCP Server Dependencies
# Core dependencies
mcp-python-sdk>=0.4.0
google-ads>=21.3.0
python-dotenv>=1.0.0
fastapi>=0.103.1
uvicorn>=0.23.2

# Async support
asyncio>=3.4.3
httpx>=0.24.1

# Utilities
pydantic>=2.3.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1 